refactor: move nestedsampler to samplers submodule
