+++ nessai/samplers/__init__.py
@@ -1,0 +1,6 @@
+# -*- coding: utf-8 -*-
+from .nestedsampler import NestedSampler
+
+__all__ = [
+    "NestedSampler",
+]
